server running, fixed naming consistency, ripped out unnessesary configurations, cleaned up, changed model naming conventions
